refactor(project): 重构项目依赖配置和实体类方法可见性
- 修改 application.yml 配置文件，添加 SQL 初始化禁用设置并移除 MyBatis 映射位置配置
- 将 mybatis-config.xml 文件从 auth 模块中移除
- 更新父 POM 版本为 2.7.18 并将 Java 版本升级到 17
- 在各模块 POM 文件中统一依赖管理，添加 Lombok、Spring Cloud 等必要依赖
- 为各模块添加 Maven 编译插件配置和 Spring Boot 插件配置
- 将多个实体类中的 pkVal 方法从 protected 改为 public 访问权限
- 在用户服务实现类中将 getLeafCount 方法返回类型从 int 改为 long
- 在主 POM 中添加项目模块声明和依赖版本属性定义

+++ module-user/src/main/java/exam/demo/moduleuser/service/impl/ResourceServiceImpl.java
@@ -82,7 +82,7 @@ public class ResourceServiceImpl extends ServiceImpl<ResourceMapper, Resource> i
         return this.listByIds(idList);
     }
 
-    private int getLeafCount(List<Resource> resources) {
+    private long getLeafCount(List<Resource> resources) {
         List<Long> parentIdList = resources.stream().map(Resource::getParentId).collect(Collectors.toList());
         QueryWrapper<Resource> wrapper = new QueryWrapper<>();
         wrapper.in(MagicPointConstant.PARENT_ID, parentIdList);
